Split the Bot class into several files
This should make things more manageable in the long run
