+++ lib/bot/_on_connect.py
@@ -1,0 +1,2 @@
+async def on_connect(self):
+    print("bot connected!")
