+++ lib/bot/_on_disconnect.py
@@ -1,0 +1,2 @@
+async def on_disconnect(self):
+    print("bot disconnected")
