+++ lib/bot/_on_message.py
@@ -1,0 +1,3 @@
+# When bot receives a message from a channel
+async def on_message(self, message):
+    pass
