+++ lib/constants.py
@@ -1,0 +1,11 @@
+# Bot command prefix
+PREFIX = "$"
+# Avoid magic numbers
+SERVER_ID = 806626416783130674
+CHANNEL_TEST = 806950823396769883
+CHANNEL_HELP_AVAILABLE = 809099330950529085
+# If the last message sent in a channel was longer than this number of minutes,
+# the idle reminder will send the reminder
+IDLE_REMINDER_MINUTES = 20
+# Every 10 minutes
+IDLE_REMINDER_CRON_TIMER = {"second": "0", "minute": "0,10,20,30,40,50"}
